+++ demo.env.js
@@ -88,4 +88,4 @@ const env = {
 
 
 // -- Export
-module.exports = env;
+export default env;